+++ recpack/pipelines/pipeline_builder.py
@@ -67,14 +67,14 @@ class PipelineBuilder(object):
 
         return arg
 
-    def add_metric(self, metric: Union[str, type], K: Union[List, int]):
+    def add_metric(self, metric: Union[str, type], K: Optional[Union[List, int]] = None):
         """Register a metric to evaluate
 
         :param metric: Metric name or type.
         :type metric: Union[str, type]
         :param K: The K value(s) used to construct metrics.
             If it is a list, for each value a metric is added.
-        :type K: Union[List, int]
+        :type K: Optional[Union[List, int]], optional
         :raises ValueError: If metric can't be resolved to a key
             in the ``METRIC_REGISTRY``.
         """
@@ -88,7 +88,8 @@ class PipelineBuilder(object):
         if isinstance(K, Iterable):
             for k in K:
                 self.add_metric(metric, k)
-        else:
+        elif K is not None:
+            # TODO Should we validate these K values to see if they make sense?
             # Check if metric already exists
             metric_name = f"{metric}_{K}"
 
@@ -96,6 +97,9 @@ class PipelineBuilder(object):
                 logger.warning(f"Metric {metric_name} already exists.")
             else:
                 self.metric_entries[metric_name] = MetricEntry(metric, K)
+        else:
+            # Bit of a hack to pass none, but it's the best I can do I think.
+            self.metric_entries[metric] = MetricEntry(metric, K)
 
     def add_algorithm(
         self,